+++ apps/ticket_web/lib/pages/home/components/ticket_cards/general_ticket_card.dart
