+++ apps/ticket_web/lib/pages/home/components/ticket_cards/personal_sponsor_ticket_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:ticket_web/gen/i18n/strings.g.dart';
 
-/// 一般チケットのカード
+/// 個人チケットのカード
 class PersonalSponsorTicketCard extends StatelessWidget {
   const PersonalSponsorTicketCard({
     required this.isLoggedIn,
